Don't attempt to delete a directory - fog will automatically clean them up (bug 100825)

+++ lib/cloud_controller/blobstore/blobstore.rb
@@ -50,7 +50,12 @@ class Blobstore
 
   def delete(key)
     blob_file = file(key)
+    return if blob_directory?(blob_file)
     blob_file.destroy if blob_file
+  end
+
+  def blob_directory?(blob_file)
+    blob_file.key.end_with?("/")
   end
 
   def download_uri(key)
